.fixed some warnings
git-svn-id: svn://ultimatepp.org/upp/trunk@6330 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/plugin/png/pngupp.cpp
@@ -231,7 +231,7 @@ bool PNGRaster::Create()
 	}
 */
 
-	data->row_bytes = png_get_rowbytes(data->png_ptr, data->info_ptr);
+	data->row_bytes = (int)png_get_rowbytes(data->png_ptr, data->info_ptr);
 	data->next_row = 0;
 	data->preload = (interlace_type != PNG_INTERLACE_NONE);
 	data->loaded = false;
